feat: enhance MarkdownToolbar with default actions and integrate into RichLexicalPostComposer for improved user experience
