+++ src/components/lexical-editor/markdown-editor.tsx
@@ -114,6 +114,7 @@ function OnChangeMarkdownPlugin({
  */
 function EditorRefPlugin({ editorRef }: { editorRef?: RefObject<LexicalEditor> }) {
   const [editor] = useLexicalComposerContext();
+  console.log("editorRef from editor ref plugin", editorRef);
 
   useEffect(() => {
     if (editorRef && "current" in editorRef) {
@@ -159,7 +160,7 @@ export function MarkdownEditor({
       <RichTextPlugin
         contentEditable={
           <ContentEditable
-            className={cn("outline-none min-h-[100px] resize-none", contentClassName)}
+            className={cn("outline-none min-h-[100px] resize-none cursor-text", contentClassName)}
           />
         }
         placeholder={<Placeholder>{placeholder}</Placeholder>}
